Refactoring the shader factory!
Former-commit-id: 8edc3725c457d1642872c7193f35b84b6c8e17a9

+++ CommonLib/DataStructures/Hashmap/Hash.h
@@ -32,7 +32,7 @@ void MurmurHash3_x86_128(const void* aKey, int aLength, uint32_t aSeed, void* ou
 
 void MurmurHash3_x64_128(const void* aKey, int aLength, uint32_t aSeed, void* out);
 
-u32 Hash(const char* key);
+u64 Hash(const char* key);
 
 //-----------------------------------------------------------------------------
